Move score calculation into Score
