+++ src/board/test/mod.rs
@@ -347,74 +347,6 @@ fn after_two_passes_the_game_should_be_over_in_tromp_taylor_rules() {
 }
 
 #[test]
-fn counting_simple_case() {
-    let mut b = Board::new(4, 6.5, Minimal);
-
-    b.play(Play(Black, 2, 1));
-    b.play(Play(White, 3, 1));
-    b.play(Play(Black, 2, 2));
-    b.play(Play(White, 3, 2));
-    b.play(Play(Black, 2, 3));
-    b.play(Play(White, 3, 3));
-    b.play(Play(Black, 2, 4));
-    b.play(Play(White, 3, 4));
-    b.play(Pass(Black));
-    b.play(Pass(White));
-
-    let (b_score, w_score) = b.score_tt();
-    assert_eq!(b_score, 8);
-    assert_eq!(w_score, 8);
-}
-
-#[test]
-fn counting_disjoint_territory() {
-    let mut b = Board::new(5, 6.5, Minimal);
-
-    b.play(Play(Black, 2, 1));
-    b.play(Play(White, 3, 1));
-    b.play(Play(Black, 2, 2));
-    b.play(Play(White, 3, 2));
-    b.play(Play(Black, 1, 3));
-    b.play(Play(White, 2, 3));
-    b.play(Play(Black, 5, 4));
-    b.play(Play(White, 1, 4));
-    b.play(Play(Black, 4, 4));
-    b.play(Play(White, 5, 3));
-    b.play(Play(Black, 4, 5));
-    b.play(Play(White, 4, 3));
-    b.play(Play(Black, 1, 2));
-    b.play(Play(White, 3, 4));
-    b.play(Pass(Black));
-    b.play(Play(White, 3, 5));
-    b.play(Pass(Black));
-    b.play(Pass(White));
-
-    let (b_score, w_score) = b.score_tt();
-    assert_eq!(b_score, 9);
-    assert_eq!(w_score, 16);
-}
-
-#[test]
-fn counting_with_neutral_points() {
-    let mut b = Board::new(5, 6.5, Minimal);
-
-    b.play(Play(Black, 2, 1));
-    b.play(Play(White, 3, 1));
-    b.play(Play(Black, 2, 2));
-    b.play(Play(White, 3, 2));
-    b.play(Play(Black, 1, 2));
-    b.play(Play(White, 2, 3));
-    b.play(Pass(Black));
-    b.play(Play(White, 1, 4));
-    b.play(Pass(Black));
-    b.play(Pass(White));
-
-    let (b_score, w_score) = b.score_tt();
-    assert_eq!(b_score, 4);
-    assert_eq!(w_score, 20);
-}
-
-#[test]
 fn capturing_two_or_more_groups_while_playing_in_an_eye_actually_captures() {
     let mut b = Board::new(5, 6.5, AnySizeTrompTaylor);
