+++ src/score/mod.rs
@@ -20,7 +20,9 @@
  ************************************************************************/
 
 use board::Black;
+use board::Board;
 use board::Color;
+use board::Coord;
 use board::Empty;
 use board::White;
 
@@ -30,17 +32,38 @@ use std::num::Float;
 
 
 pub struct Score {
-    color: Color,
-    score: f32
+    black_stones: usize,
+    color:        Color,
+    score:        f32,
+    white_stones: usize,
+}
+
+struct Territory {
+    color:  Color,
+    coords: Vec<Coord>,
+}
+
+impl Territory {
+
+    pub fn new() -> Territory {
+        Territory { color: Empty, coords: Vec::new() }
+    }
 }
 
 mod test;
 
 impl Score {
-    pub fn new(scores: (usize, usize), komi: f32) -> Score {
-        let (bs, ws) = scores;
+
+    // Figure out which methods board needs to provide for this to
+    // work so that we can create a trait with just those methods for
+    // our test purposes.
+    //
+    // Store a reference to the Board in Score and compute the score
+    // in an instance method.
+    pub fn new(board: &Board) -> Score {
+        let (bs, ws) = Score::score_tt(board);
         let b_score = bs as f32;
-        let w_score = (ws as f32) + komi;
+        let w_score = (ws as f32) + board.komi();
         let score = (b_score - w_score).abs();
         let color = if b_score == w_score {
             Empty
@@ -49,11 +72,75 @@ impl Score {
         } else {
             White
         };
-        Score {color: color, score: score}
+        Score {color: color, score: score, white_stones: ws, black_stones: bs}
+    }
+
+    // TODO: Make this private
+    pub fn score_tt(board: &Board) -> (usize, usize) {
+        let mut black_score = 0;
+        let mut white_score = 0;
+        for point in board.points().iter() {
+            match point.color {
+                Black => { black_score += 1; },
+                Empty => {},
+                White => { white_score += 1; },
+            }
+        }
+        let mut empty_intersections = board.vacant().clone();
+        while empty_intersections.len() > 0 {
+            let territory = Score::build_territory_chain(empty_intersections[0], board);
+
+            match territory.color {
+                Black => black_score += territory.coords.len(),
+                White => white_score += territory.coords.len(),
+                Empty => () // This territory is not enclosed by a single color
+            }
+
+            empty_intersections = empty_intersections.into_iter().filter(|coord| !territory.coords.contains(coord)).collect();
+        }
+        (black_score, white_score)
+    }
+
+    fn build_territory_chain(first_intersection: Coord, board: &Board) -> Territory {
+        let mut territory_chain = Territory::new();
+        let mut to_visit = Vec::new();
+        let mut neutral  = false;
+
+        to_visit.push(first_intersection);
+
+        while to_visit.len() > 0 {
+            let current_coord = to_visit.pop().unwrap();
+            if !territory_chain.coords.contains(&current_coord) {territory_chain.coords.push(current_coord);}
+
+            for &coord in board.neighbours(current_coord).iter() {
+                match board.color(&coord) {
+                    Empty => if !territory_chain.coords.contains(&coord) {to_visit.push(coord)},
+                    col   => if territory_chain.color != Empty && territory_chain.color != col {
+                        neutral = true;
+                    } else {
+                        territory_chain.color = col;
+                    }
+                }
+            }
+        }
+
+        if neutral {
+            territory_chain.color = Empty;
+        }
+
+        territory_chain
     }
 
     pub fn color(&self) -> Color {
         self.color
+    }
+
+    pub fn white_stones(&self) -> usize {
+        self.white_stones
+    }
+
+    pub fn black_stones(&self) -> usize {
+        self.black_stones
     }
 }
